Algumas mudanças
Aplicando importações e refatorando código.

+++ SimuladorColisoesEsferas/src/Vetor.java
@@ -1,20 +1,20 @@
 public class Vetor {
-    private int x;
-    private int y;
+    private long x;
+    private long y;
 
-    public int getX() {
+    public long getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(long x) {
         this.x = x;
     }
 
-    public int getY() {
+    public long getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(long y) {
         this.y = y;
     }
 }
